if length != 0, run

+++ wallpapers.rb
@@ -14,10 +14,12 @@ def get_new_wallpapers(pic_array, min_width)
     # If photo width is larger the min_width (screen width), put it in an array to be saved
     to_write << pic if (pic['width'] >= min_width) && (pic['width'] > pic['height']) && (pic['height'] > min_width * 1.6)
   end
-  # Delete current photos
-  Dir.foreach('wallpapers/') {|f| fn = File.join('wallpapers/', f); File.delete(fn) if f != '.' && f != '..'}
-  # Save the new photos
-  to_write.each { |pic| File.write("wallpapers/#{pic['id']}.jpg", open(pic['image_url']).read) }
+  if to_write.length != 0
+    # Delete current photos
+    Dir.foreach('wallpapers/') {|f| fn = File.join('wallpapers/', f); File.delete(fn) if f != '.' && f != '..'}
+    # Save the new photos
+    to_write.each { |pic| File.write("wallpapers/#{pic['id']}.jpg", open(pic['image_url']).read) }
+  end
 end
 
 # Make call to get response, parse to JSON, strip to just photo array
